backend/app/generator.py
import os
import uuid
import time
from datetime import datetime
import numpy as np
from PIL import Image
import io
import base64

from .config import settings

class Model3DGenerator:
    def __init__(self):
        # Здесь загружаем вашу модель
        print(f"Подключение к  сервису с моделью")
        
    def generate_from_prompt(self, prompt, style="realistic", quality=75):
        """Генерация 3D-модели из текстового промпта"""
        print(f"Генерация модели для промпта: '{prompt}'")
        
        # Здесь ваша логика генерации
        # Это заглушка, которая создает тестовый файл GLB
        time.sleep(1.5)  # Имитация времени генерации
        
        # Создаем уникальный ID и путь
        model_id = str(uuid.uuid4())
        filename = f"{model_id}.glb"
        filepath = os.path.join(settings.MODELS_DIR, filename)
        
        # Здесь должна быть ваша логика генерации модели
        # Для примера создадим пустой файл
        with open(filepath, "wb") as f:
            # В реальном проекте здесь будет бинарные данные вашей модели
            f.write(b"Fake GLB file content for demonstration")
        
        return {
            "model_id": model_id,
            "filepath": filepath,
            "format": "glb",
            "size": os.path.getsize(filepath)
        }
    
    def generate_from_image(self, image_data, prompt=None, style="realistic"):
        """Генерация 3D-модели из изображения"""
        # Декодируем изображение
        image = self._decode_image(image_data)
        
        # Здесь ваша логика обработки изображения
        print("Обработка изображения для генерации 3D-модели")
        
        return self.generate_from_prompt(prompt or "Generated from image", style)
    
    def _decode_image(self, base64_data):
        """Декодирование base64 изображения"""
        # Удаляем префикс data:image/png;base64,
        if ',' in base64_data:
            base64_data = base64_data.split(',')[1]
        
        image_data = base64.b64decode(base64_data)
        return Image.open(io.BytesIO(image_data))

# Создаем экземпляр генератора
generator = Model3DGenerator()