backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
numpy
Pillow